+++ src/main/java/com/notenoughmail/kubejs_tfc/util/implementation/mixin/TFCChunkGeneratorMixin.java
@@ -1,33 +1,21 @@
 package com.notenoughmail.kubejs_tfc.util.implementation.mixin;
 
-import com.mojang.serialization.JsonOps;
-import com.notenoughmail.kubejs_tfc.KubeJSTFC;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.mojang.serialization.MapCodec;
 import com.notenoughmail.kubejs_tfc.event.ModifyDefaultWorldGenSettingsEventJS;
-import com.notenoughmail.kubejs_tfc.util.EventHandlers;
-import com.notenoughmail.kubejs_tfc.util.WorldGenUtils;
 import net.dries007.tfc.world.TFCChunkGenerator;
-import net.dries007.tfc.world.biome.BiomeSourceExtension;
 import net.dries007.tfc.world.settings.Settings;
-import net.minecraft.core.Holder;
-import net.minecraft.world.level.levelgen.NoiseGeneratorSettings;
+import org.objectweb.asm.Opcodes;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(value = TFCChunkGenerator.class, remap = false)
 public abstract class TFCChunkGeneratorMixin {
 
-    @Shadow(remap = false)
-    private Settings settings;
-
-    @Inject(method = "<init>", at = @At("TAIL"), remap = false)
-    private void kubejs_tfc$ModifyDefaultSettings(BiomeSourceExtension arg0, Holder<NoiseGeneratorSettings> arg1, Settings arg2, CallbackInfo ci) {
-        if (!WorldGenUtils.worldgenHasBeenTransformed && EventHandlers.defaultSettings.hasListeners()) {
-            EventHandlers.defaultSettings.post(new ModifyDefaultWorldGenSettingsEventJS((TFCChunkGenerator) (Object) this));
-            KubeJSTFC.warningLog("Modified worldgen settings: {}", () -> Settings.CODEC.encoder().encodeStart(JsonOps.INSTANCE, settings).get().left().get());
-            WorldGenUtils.worldgenHasBeenTransformed = true;
-        }
+    // TODO: Verify the lambda name is consistent, verify the #mapResult does not change the class & causes issues somewhere else
+    @WrapOperation(method = "lambda$static$3", at = @At(value = "FIELD", target = "Lnet/dries007/tfc/world/settings/Settings;CODEC:Lcom/mojang/serialization/MapCodec;", opcode = Opcodes.GETSTATIC), remap = false)
+    private static MapCodec<Settings> kubejs_tfc$Modify(Operation<MapCodec<Settings>> getField) {
+        return getField.call().mapResult(ModifyDefaultWorldGenSettingsEventJS.SETTINGS_TRANSFORMER);
     }
 }